refactor: restructure analytics module and utils for better organization
- Split large analytics.ts file into modular structure:
  * counts.ts - counting functions
  * groups.ts - grouping functions
  * statistics.ts - statistical calculations
  * participations.ts - participation tracking
  * types.ts - analytics-specific types
- Reorganize utils into separate modules:
  * array.ts - array utility functions
  * date.ts - date/time utility functions
- Update test structure to match new modular organization
- Improve code maintainability and separation of concerns
