+++ src/core/utils/array.ts
@@ -1,0 +1,3 @@
+export function uniqueArray<T>(array: T[] | null): T[] | null {
+  return array ? Array.from(new Set(array)) : null;
+}
