+++ src/core/utils/index.ts
@@ -1,0 +1,2 @@
+export * from './array';
+export * from './date';
